LayoutTests/storage/indexeddb/deletedatabase-blocked-expected.txt

Test that deleteDatabase is blocked on open connections

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

dbname = "deletedatabase-blocked.html"
blockedEventFired = false
indexedDB.open(dbname)

openOnSuccess():
h = event.target.result
indexedDB.deleteDatabase(dbname)

onVersionChange():
old = 1
new = ""
scheduling timeout to close

deleteDatabaseOnBlocked():
blockedEventFired = true

timeoutCallback():
h.close()

deleteDatabaseOnSuccess():
PASS blockedEventFired is true
PASS successfullyParsed is true

TEST COMPLETE
